Basic command line unit tests
