+++ orgnetsim/serve_test.go
@@ -5,42 +5,22 @@ import (
 	"testing"
 )
 
-func IsFalse(t *testing.T, condition bool, msg string) {
-	if condition {
-		t.Error(msg)
-	}
-}
-
-func IsTrue(t *testing.T, condition bool, msg string) {
-	if !condition {
-		t.Error(msg)
-	}
-}
-
-func AreEqual(t *testing.T, expected interface{}, actual interface{}, msg string) {
-	if expected != actual {
-		t.Errorf("%s Expected = %v Actual = %v", msg, expected, actual)
-	}
-}
-
-func NotEqual(t *testing.T, expected interface{}, actual interface{}, msg string) {
-	if expected == actual {
-		t.Errorf("%s Expected = %v Actual = %v", msg, expected, actual)
-	}
-}
-
-func AssertSuccess(t *testing.T, err error) {
-	if err != nil {
-		t.Errorf(err.Error())
-	}
-}
-
 func TestServeReturnsFalseForHelp(t *testing.T) {
 	oldArgs := os.Args
 	defer func() { os.Args = oldArgs }()
 	os.Args = []string{"orgnetsim", "serve", "-help"}
 	success, staticDir, so := serveCommandLineOptions()
 	IsFalse(t, success, "-help not returning false")
+	AreEqual(t, so.Port, "8080", "Incorrect default port")
+	AreEqual(t, staticDir, "", "Default staticDir should be empty")
+}
+
+func TestServeReturnsTrue(t *testing.T) {
+	oldArgs := os.Args
+	defer func() { os.Args = oldArgs }()
+	os.Args = []string{"orgnetsim", "serve", "tmpDir"}
+	success, staticDir, so := serveCommandLineOptions()
+	IsTrue(t, success, "not returning true")
 	AreEqual(t, so.Port, "8080", "Incorrect default port")
 	AreEqual(t, staticDir, "", "Default staticDir should be empty")
 }
